Team Admin functionality

+++ app/Http/Controllers/Admin/UsersController.php
@@ -17,7 +17,7 @@ class UsersController extends Controller
 {
     public function index()
     {
-        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('user_access') || !auth()->user()->is_team_admin, Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $users = User::all();
 
@@ -26,7 +26,7 @@ class UsersController extends Controller
 
     public function create()
     {
-        abort_if(Gate::denies('user_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('user_create') || !auth()->user()->is_team_admin, Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $roles = Role::all()->pluck('title', 'id');
 
@@ -37,6 +37,8 @@ class UsersController extends Controller
 
     public function store(StoreUserRequest $request)
     {
+        abort_if(!auth()->user()->is_team_admin, Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $user = User::create($request->all());
         $user->roles()->sync($request->input('roles', []));
 
@@ -46,7 +48,7 @@ class UsersController extends Controller
 
     public function edit(User $user)
     {
-        abort_if(Gate::denies('user_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('user_edit') || !auth()->user()->is_team_admin, Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $roles = Role::all()->pluck('title', 'id');
 
@@ -59,6 +61,8 @@ class UsersController extends Controller
 
     public function update(UpdateUserRequest $request, User $user)
     {
+        abort_if(!auth()->user()->is_team_admin, Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $user->update($request->all());
         $user->roles()->sync($request->input('roles', []));
 
@@ -68,7 +72,7 @@ class UsersController extends Controller
 
     public function show(User $user)
     {
-        abort_if(Gate::denies('user_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('user_show') || !auth()->user()->is_team_admin, Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $user->load('roles', 'team');
 
@@ -77,7 +81,7 @@ class UsersController extends Controller
 
     public function destroy(User $user)
     {
-        abort_if(Gate::denies('user_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('user_delete') || !auth()->user()->is_team_admin, Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $user->delete();
 
@@ -87,6 +91,8 @@ class UsersController extends Controller
 
     public function massDestroy(MassDestroyUserRequest $request)
     {
+        abort_if(!auth()->user()->is_team_admin, Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         User::whereIn('id', request('ids'))->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
